Host SVG icons directly instead of using fontawesome
Why
===

The main reason is size. The rustsec.org home page, before this commit was
applied, was 152K (compressed). 75K of that was fontawesome-webfont.woff,
and rustsec only uses four icons from there. It also cuts a DNS lookup.

See also: https://www.jefftk.com/p/shared-cache-is-going-away

The second reason was accessibility. I run with webfonts turned off by
default, because they're usually not important enough to be worth waiting
for, but occasionally a site (like rustsec.org) relies on them to understand
the buttons. Other people can't see them because they're blind (notice that
this patch also adds aria-label and title attributes), or dyslexic. We should
not stand for accessibility failures in Rust's web properties.

See also: https://github.blog/2016-02-22-delivering-octicons-with-svg/

How
===

To keep the page nice and snappy, fontawesome icons are `{% include %}`-ed
directly into the page. This is the same way GitHub went about removing
their use of icon fonts, and it has the same advantages (reduced jank,
and the ability to recolor the icon with CSS). Luckily, none of these icons
are very complicated.